Started working on new random excitation generators
--> As discussed with asa / ghb
--> Use fewer random numbers, and be somewhat more uniform

+++ src/FciMCData.F90
@@ -118,7 +118,7 @@ MODULE FciMCData
 
       ! The approximate fraction of singles and doubles. This is calculated
       ! using the HF determinant, if using non-uniform random excitations.
-      real*8 :: pDoubles, pSingles
+      real(dp) :: pDoubles, pSingles
       
       ! Bit representation of the HF determinant
       integer(kind=n_int), allocatable :: iLutHF(:)
